requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
google-cloud-language==2.12.0
google-cloud-core==2.3.3
redis==5.0.1
geopy==2.4.0
python-multipart==0.0.6
